evaluate static string matches in matchless
